Removed progress link.
The progress link can only be viewed by someone who is either yourself (while logged in) or by someone who has you added as a friend on the projecteuler.net website. 

I guess it doesn't have to be removed from the README, but at the very least I thought it was worth bringing up; especially since the point of your GitHub is to put the source code here anyway. ;)